+++ Tests/Runtime/ClientSendTests.cs
@@ -1,4 +1,5 @@
 ﻿using Backtrace.Unity.Model;
+using Backtrace.Unity.Model.JsonData;
 using NUnit.Framework;
 using System;
 using System.Collections;
@@ -67,6 +68,133 @@ namespace Backtrace.Unity.Tests.Runtime
             {
                 trigger = true;
             });
+            yield return new WaitForEndOfFrame();
+            Assert.IsTrue(trigger);
+            yield return null;
+        }
+
+        [UnityTest]
+        public IEnumerator PiiTests_ShouldRemoveEnvironmentVariables_AnnotationsShouldntBeAvailable()
+        {
+            var trigger = false;
+            var exception = new Exception("custom exception message");
+            client.BeforeSend = (BacktraceData data) =>
+             {
+                 Assert.IsNotNull(data.Annotation.EnvironmentVariables);
+                 data.Annotation.EnvironmentVariables = null;
+                 return data;
+             };
+
+            client.RequestHandler = (string url, BacktraceData data) =>
+            {
+                trigger = true;
+                Assert.IsNull(data.Annotation.EnvironmentVariables);
+                return new BacktraceResult();
+            };
+            client.Send(exception);
+
+            yield return new WaitForEndOfFrame();
+            Assert.IsTrue(trigger);
+            yield return null;
+        }
+
+        [UnityTest]
+        public IEnumerator PiiTests_ShouldChangeApplicationDataPath_ApplicationDataPathDoesntHaveUserNameAnymore()
+        {
+            var trigger = false;
+            var exception = new Exception("custom exception message");
+            var expectedDataPath = "/some/path";
+            client.BeforeSend = (BacktraceData data) =>
+            {
+                Assert.IsNotNull(data.Attributes.Attributes["application.data_path"]);
+                data.Attributes.Attributes["application.data_path"] = expectedDataPath;
+                return data;
+            };
+            client.RequestHandler = (string url, BacktraceData data) =>
+            {
+                trigger = true;
+                Assert.AreEqual(expectedDataPath, data.Attributes.Attributes["application.data_path"]);
+                return new BacktraceResult();
+            };
+            client.Send(exception);
+
+            yield return new WaitForEndOfFrame();
+            Assert.IsTrue(trigger);
+            yield return null;
+        }
+
+        [UnityTest]
+        public IEnumerator PiiTests_ShouldChangeSourceCodeIntegration_SourceCodeTextDoesntHaveUserNameAnymore()
+        {
+            var trigger = false;
+            var exception = new Exception("custom exception message");
+            var sourceCodeTestString = "source-code-test-string";
+            client.BeforeSend = (BacktraceData data) =>
+            {
+                Assert.IsNotNull(data.SourceCode);
+                Assert.IsNotEmpty(data.SourceCode.Text);
+                data.SourceCode.Text = sourceCodeTestString;
+                return data;
+            };
+            client.RequestHandler = (string url, BacktraceData data) =>
+            {
+                trigger = true;
+                Assert.AreEqual(sourceCodeTestString, data.SourceCode.Text);
+                return new BacktraceResult();
+            };
+            client.Send(exception);
+
+            yield return new WaitForEndOfFrame();
+            Assert.IsTrue(trigger);
+            yield return null;
+        }
+
+        [UnityTest]
+        public IEnumerator PiiTests_ShouldModifyEnvironmentVariable_IntegrationShouldUseModifiedEnvironmentVariables()
+        {
+            var trigger = false;
+            var exception = new Exception("custom exception message");
+
+            var environmentVariableKey = "foo";
+            var expectedValue = "bar";
+            Annotations.EnvironmentVariablesCache[environmentVariableKey] = expectedValue;
+
+            client.BeforeSend = (BacktraceData data) =>
+            {
+                var actualValue = data.Annotation.EnvironmentVariables[environmentVariableKey];
+                Assert.AreEqual(expectedValue, actualValue);
+                trigger = true;
+                return data;
+            };
+            client.Send(exception);
+
+            yield return new WaitForEndOfFrame();
+            Assert.IsTrue(trigger);
+            yield return null;
+        }
+
+
+        [UnityTest]
+        public IEnumerator PiiTests_ShouldRemoveEnvironmentVariableValue_IntegrationShouldUseModifiedEnvironmentVariables()
+        {
+            var trigger = false;
+            var exception = new Exception("custom exception message");
+
+            var environmentVariableKey = "USERNAME";
+            var expectedValue = "%USERNAME%";
+            var defaultUserName = Annotations.EnvironmentVariablesCache[environmentVariableKey];
+            Annotations.EnvironmentVariablesCache[environmentVariableKey] = expectedValue;
+
+            client.BeforeSend = (BacktraceData data) =>
+            {
+                var actualValue = data.Annotation.EnvironmentVariables[environmentVariableKey];
+                Assert.AreEqual(expectedValue, actualValue);
+                Assert.AreNotEqual(defaultUserName, actualValue);
+                trigger = true;
+                return data;
+            };
+            client.Send(exception);
+
             yield return new WaitForEndOfFrame();
             Assert.IsTrue(trigger);
             yield return null;